requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
mysqlclient==2.2.4
pykg-config==1.3.0
pytz==2024.1
sqlparse==0.4.4
